+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v4.2.0
 
 - star threads
+- also hide comments of blocked users in quotes
+- add reason to block
 
 ## v4.0.3